LPS-150119 Fix selecting support seat roles on dropdown
